Added UI support to create new bridge session.
A menu 'New Session' is added under 'File' menu, and it contains all the bridges available
to ModelBuilder, and will be updated when a new bridge is loaded through loading plugin.
So now user can start a new bridge session to load or create models through that bridge.
The use case here is that bridges sometimes can offer operation to create model from scratch,
for example, creating sphere or prism from cgm-open cascase. Also with this commit, the
application will listen to the operation-signal emitted from the model tree view to start
the operation, and process the result as desired.