Fixed a bug with reset camera when all entities are off.
In paraview, when all blocks are turned off individually from render view,
the representation of the whole dataset is still considered visible, which
lead to the reset camera still use its bounds for calculation. A check is
added so that when all blocked are turned off, the rep is set invisible.