Fixed the duplicate-command-function error while starting up
When starting ModelBuilder, there are a couple plugins loaded automatically, and
they bring in same set of dependent libraries, which vtkClientServerInterperter
will see as different libraries with same function symbol, thus error out. The
fix is to remove the pv-plugin CMake variable before calling add_paraview_plugin for
one of the plugins so that its initialization will not be contaminated from previous plugins.