Added colorby toolbar in main window.
This toolbar is linked with the active representation in the render view, which will be changed by mouse click, or select in the model tree view. The toolbar is the same as the one in paraview. Also in this commit, a bunch of unused headers and legacy routines were cleaned up.