Refactored ColorBy toolbar in ModelBuilder main window
When assigning a color in the tree view, all children of the assigned index got the same color, which should not be the case. Only assigned index (entity) should get the assigned color, and according to what the current selection in "ColorBy" toolbar, the entity representation should only displays with that color, if the ColorBy selection matches its entity type. Ohter entities will be colored through color transfter function.