Add separate LUT for individual model representation.
In CMB, we want to color different models with their own LUT, and these LUTs use the fielddata
array with the same name. For example, color a model by Volume, Entity, or Group field data.
In paraview however, the CTF manager will use the same LUT if it sees the same ColorArrayName
across all representations, therefore when we try to color a different model with a fielddata array with the same name, the same LUT will be used for both models, which will not work for CMB.