Added a feature to view model geometry by assigned attributes.
This commit includes UI components and application logic to color the model with associated attributes. To achieve this, a new filter is added for manipulating the field data array of multiblock model source. Due to a bug in paraview when the field data array used for coloring a multiblock only partillay exists on some blocks, there will be color leaks to other blocks, the work-around in this case is to add dummy field data array to the blocks that don't have valid data. Also in this commit, the logic for coloring-by group is also changed to use the new filter.