Added support for loading optional autoload plugins in ModelBuilder
Added a couple cmake variables to allow developer specify plugins to
autoload into ModelBuilder when launched. A new cmake varialble
CMB_OPTIONAL_PLUGINS is a semicolon separated string list of plugin names,
and another new variable CMB_TEST_PLUGIN_PATHS is a semicolon separated
list of directories for those plugins. A .plugins file will be created,
copied and installed to proper locations for both build tree and install
tree so that these plugins will be auto-loaded into ModelBuilder when launched.
This feature is necessary for cmb tests with ctest because when running tests
from ctest and we ignore the registry info so that the tests can be run cleanly, but
this also means the auto-load-plugins in registry will not be loaded, and ModelBuilder
relies on session plugins to load models from different bridges.