Update vtkPVRenderViewSettings default to OFFSET_FACES.
Paraview and VTK changed how coincident geometries are handled, which broke CMB's use of model
vertex, edge and face. In CMB, we expect the rendering order is face, edge, vertex, with vertex
always on top of edge, edge always on top of face. This commit fixes vertex and edge are always
on top of face, but we still need to figure out how to render vertex always on top of edge.