Fixed a couple issues with loading different simulation files.
If there is a simulation file already loaded, loading another simulation file will
cause the model entites not showing up in the attribute-entity association window.
Also, it will crash in certain cases where model manager is directly used in attribute
system to query entities. The bug is that when loading a different file, the application
would clear current attribute system, and create a new attribute system, but did not set
the model manager to the new attribute system, hence the issues.