Fixes to create session with sessionId while importing smtk file
While importing a smtk file, there is session name and id in the smtk file, so when creating a session for importing the file, if the sessionId in the file has corresponding existing session available, just use that session, or create a new session using the sessionId. This way only one session is used or created for importing.