Fix some selection issues on models with meshes
When models and meshes are selected in the tree view, separate signals are emitted, and the application will process the signals individually. If the application uses paraview's pqSelectionManager::select() to select therepresentation, it will clear what's already selected first, this behavior causes different model and mesh representations can't be selected at the same time. The fix is to just set the selection input to the sources, then call view->render() to update all selections in render view.