Fixed a performance bug in creating arc widget representation
There were some logic flaws in creating arc widget representation from a polydata,
which caused a N-square algorithm of re-building lines for contour representation.
After fixing the logic, the time building contour is shorterned by a factor of almost 200,
which makes it possible to handle a whole arc with 20k+ points with ease.