Added logic to handle features that need active model representation
In certain cases, such as when a reader is created in paraview, but there is no data representation is created for that reader, for example, a reader for template file, or an internal smproxy source, the pqActiveObjects will set its activePort to the new source, which will lead to no active representation in pqActiveObjects, which certain features are relying on, such as color-by or selection operation from sessions. New logic is added to make sure that if model(s) are selected, or there are models in the system, we will activate one model when asked for active model. This commit also links the tree view selection to activation of models selected.