Add support for load vtp points data in PointsBuilder
Added support in PointsBuilder to load in points cloud in vtp format. However, the reading will not be controlled by number of targe points or ROI, because the underneath reader, vtkXMLPolyDataReader doesn't support them. A warning message stating the above limitation will be presented to the user for vtp files.