Link image selection in render view to tree view
Now selecting image_url entry in the tree view will make the image representation active, meaning the Display properties panel will be linked to it. By default the image representation will be set to Slice, which can not be selected in render view due to lack-of-image-render-pass in Paraview-VTK selection mechanism. However if the image representation is set to other types, such as Surface, then the image can be picked in render view (be ware of the initial slow render for generating surface geometry), and the image_url entry in the tree view should be highlighted.