Add measurement capabilty to CMB
Add Ruler feature to CMB based on the Ruler in Paraview to do measurements.
There are two parts of ruler in PV as rulerWidget and vtkDistanceRepresentation2D.
The later one is not added for now because CMB model builder has no
panel for sources, and we do not need to add a line source after
measurement. An XML test is also added.
Close issue #99.