Fixed a bug in resetting UI while reloading a simulation template
When reloading a template, or loading another tempate after a template is loaded, the SimBuiler UI will be reset for the new template. The bug is that the reference to the model tree panel was not reset for the new UI, which broke the linkage between the entity-list-selection on the attribute UI and the actual selection of model tree.