Fix mesh panel after closing a model.
The mesh panel holds on to the model manager pointer it is given
the first time it is created/displayed. However, the model manager
is destroyed and recreated on the "close" model action (Ctrl-W/Cmd-W)
and then re-instantiated by the pqCMBModelManager. Signal when
this happens and tell the panel to update its qtRemusMesherSelector.

This should close issue #76.