Fix selection for volumes and groups when in entity list view
When switching to entity list view type on model panel, the volume or gorup selection is not being linked with render view. To fix this, more logic is added to handle volume and group selections on the model panel. Also in this commit, pqCMBContextMenuHelper.h file is converted to a class with static member functions. Lastly some minor code clean up.