Add application logic to handle separate Auxiliary representation
For auxiliary geometry entities, they could be displayed as part of the associated model or a completely separate representation from the model. The idea behind it is so that the display properties, such as coloring scheme, and positions can be controlled separately from model representation. This is desired in certain use cases. The selection and setting-property logic (visibility and color) are also modified to accomodate these use cases.