ENH: Changed Camera Interaction Mechanism
The old Combox box has been replaced with a pair of actions.  The main functionality has moved to the
main wondow class from the main window core class.  The methods include:

1. Ability to change camera interaction 2D/3D
2. Ability to pop and push the current camera interaction
3. Ability to lock and unlock the camera interaction UI

There are a twin set of methods added to the Main Window Core Class.  These methods
allow the camera inetraction to be accessed for the core - these methods send signals
to the Main Window class that performs the appropriate calls.