BUG: Fixed selection issue due to Close Data action
The problem was when the Close Data Action was triggered you could no
longer select model entities.  The cause was due to the selection manager
not being told there was a new client side SMTK model manager (it was
\holding onto the now empty old manager)

The fix was to notify the selection manager when the new model
manager was created.

Also - changed the slot signature to remove the proxy being passed in
since it was never used.