Switch interaction modes earlier.
Now we switch interaction modes both when a model
is created (at which point the dimension may not be
known if the model is empty) **and** when the first
cell with a tessellation is added.

Although currently ignored, we also emit a signal
each time one or more auxiliary geometry instances
is created so that the interaction mode might also
respond to new imaages, etc.