BUG: Fix selection color not shown when entities have assigned colors
Selections are not drawn properly after color assignment if
both the selection and original representation are rendered since PV's
rendering paints in the order in which representations are modified.
Turn off selected blocks in the original representation to avoid
Z-fighting and overdraw.

When update entities color after operation, CMB would turn off
selEntities block visiblity as well. If a new selection is triggered,
selEntities would be turned on for selection.